ui add images to correct folders [#17]
